+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * ADD: (Beta) Async support to RuuviTagReactive
 * FIX: (Beta) Use passive scanning on macOS with Bleak
 * ADD: (Beta) Async version to RuuviTag-class
+* ADD: (Beta) Async get_data_for_sensors_async method
 
 ## [2.1.0] - 2023-02-18
 * CHANGE: Add RSSI reporting